requirements.txt
pygame==2.5.2
python-rtmidi==1.5.8
numpy==1.24.3
pillow==10.0.0